+++ beta-src/src/utils/state/gameApiSlice/extraReducers/fetchGameStatus/fulfilled.ts
@@ -2,7 +2,7 @@ import { current } from "@reduxjs/toolkit";
 import GameDataResponse from "../../../../../state/interfaces/GameDataResponse";
 import GameOverviewResponse from "../../../../../state/interfaces/GameOverviewResponse";
 import { GameState } from "../../../../../state/interfaces/GameState";
-import { handleGetSucceeded, handleGetFailed } from "../handleFulfillReject";
+import { handleGetSucceeded, handleGetFailed } from "../handleSucceededFailed";
 
 /* eslint-disable no-param-reassign */
 export default function fetchGameStatusFulfilled(
